requirements.txt
flask>=3.0.0
twilio>=9.0.0
google-cloud-vision>=3.7.0
google-cloud-storage>=2.16.0
requests>=2.31.0
gunicorn>=21.2.0
google-generativeai>=0.7.0
duckdb>=0.10.0
pandas>=2.2.0
tabulate>=0.9.0
torch>=2.0.0
transformers>=4.40.0
scikit-learn>=1.4.0
rapidfuzz>=3.0.0
anthropic>=0.25.0
streamlit>=1.35.0
numpy>=1.26.0
plotly>=5.20.0
huggingface-hub
tokenizers
safetensors
opencv-python
Pillow
pydantic
python-dateutil
regex
matplotlib
pytest
groq
python-dotenv
